requirements.txt
amqp==5.0.5
appdirs==1.4.3
asn1crypto==1.4.0
awscli==1.19.61
Babel==2.9.0
bcrypt==3.2.0
beautifulsoup4==4.6.0
billiard==3.6.3.0
blis==0.7.4
boto3==1.17.62
botocore==1.20.62
cachetools==4.2.1
catalogue==2.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coloredlogs==15.0
contextvars==2.4
coverage==4.4.2
cryptography==2.1.4
cssselect==1.1.0
cymem==2.0.5
dataclasses==0.8
decorator==4.0.9
defusedxml==0.7.0
Django==2.2
django-allauth==0.36.0
django-bulk-update==2.2.0
django-constance==2.0.0
django-filter==2.4.0
django-jsonview==2.0.0
django-longerusernameandemail==0.5.7
django-mobile==0.7.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-paypal==1.1
django-picklefield==3.0.1
django-queryinspect==1.0.0
django-redis-cache==2.1.1
django-storages==1.6.5
django-user-sessions==1.7.1
django-watchman==0.15.0
djangorestframework==3.12.2
docutils==0.15.2
drf-nested-routers==0.93.3
dropbox==11.3.0
ecdsa==0.11
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.1
Fabric==1.14.0
falcon==2.0.0
falcon-cors==1.1.7
falcon-multipart==0.2.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
flake8==2.6.2
Flask==1.1.2
flower==0.9.7
freezegun==0.3.6
funcsigs==1.0.2
gevent==21.1.2
google-api-python-client==1.7.12
google-auth==1.27.1
google-auth-httplib2==0.1.0
greenlet==1.0.0
gunicorn==19.6.0
html5lib==0.999
httplib2==0.18.1
idna==2.10
img2pdf==0.4.0
immutables==0.15
importlib-metadata==3.7.0
inflect==0.2.5
inflection==0.3.1
ipaddress==1.0.16
ipython-genutils==0.2.0
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonfield==3.1.0
kombu==5.0.2
langdetect==1.0.5
locustio==0.7.3
lxml==4.6.3
marisa-trie==0.7.5
MarkupSafe==1.1.1
mccabe==0.5.3
mock==2.0.0
model-bakery==1.2.1
model-mommy==1.3.2
msgpack-python==0.5.6
murmurhash==1.0.5
ndg-httpsclient==0.4.0
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.0
oauth2client==4.1.3
oauthlib==2.0.2
ocrmypdf==11.7.3
packaging==20.9
paramiko==2.4.3
pathy==0.4.0
pbr==3.0.1
pdf2docx==0.5.1
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==2.11.1
Pillow==8.1.1
pluggy==0.13.1
ply==3.11
preshed==3.0.5
prompt-toolkit==3.0.16
psycopg2==2.8
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.6.2
pycodestyle==2.0.0
pycparser==2.20
pycrypto==2.6.1
pydantic==1.7.3
PyDrive==1.3.1
pyflakes==1.2.3
Pygments==2.8.1
PyMuPDF==1.18.12
PyNaCl==1.4.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
PyReadableDiff==0.2.3
PyStemmer==2.0.1
python-dateutil==2.6.0
python-docx==0.8.10
python-dotenv==0.17.1
python-http-client==2.2.1
python-intercom==2.1.1
python-magic==0.4.13
python-mimeparse==1.6.0
python-openid==2.2.5
python-redis-lock==2.3.0
python3-openid==3.2.0
pytz==2016.10
PyYAML==5.4.1
redis==3.2.0
regex==2021.3.17
reportlab==3.5.67
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==0.8.0
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.5.0
selenium==2.53.6
sendgrid==3.6.3
sendgrid-django==4.0.4
sgmllib3k==1.0.0
simplegeneric==0.8.1
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
South==1.0.2
spacy==3.0.3
spacy-legacy==3.0.1
sqlparse==0.4.1
srsly==2.4.0
stone==3.2.1
swapper==1.1.2.post1
termcolor==1.1.0
thinc==8.0.1
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
typer==0.3.2
typing-extensions==3.7.4.3
Unidecode==0.4.16
uritemplate==3.0.1
urllib3==1.26.3
validators==0.10.1
vine==5.0.0
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==1.0.1
wikipedia==1.4.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
django-cors-headers==3.7.0